+++ tests/format/angular/control-flow/if.html
@@ -25,3 +25,9 @@
 
 @if      (foo(     "quotes")) {}
 @else if (foo(     'quotes')) {}
+
+
+@if (widthCategory !== undefined &&
+widthCategory > topStartToSideStartMaxSize) {
+<app-component />
+}
